requirements.txt
requests==2.24.0
plotly==4.12.0
python-kasa==0.4.0.dev1
tinydb==4.3.0
python-telegram-bot==13.1
paho-mqtt==1.5.1